fix(cf-deploy-config-writer): refactor hasUI5CliV3 to project-access for common reuse
