+++ .changeset/dull-points-nail.md
@@ -1,0 +1,6 @@
+---
+'@sap-ux/odata-service-writer': patch
+'@sap-ux/project-access': patch
+---
+
+Move hasUI5CliV3 to project-access for common re-use
