+++ packages/project-access/test/project/script.test.ts
@@ -1,7 +1,7 @@
 import { join } from 'path';
 import { create as createStorage } from 'mem-fs';
 import { create } from 'mem-fs-editor';
-import { FileName, updatePackageScript } from '../../src';
+import { FileName, updatePackageScript, hasUI5CliV3 } from '../../src';
 
 describe('Test updatePackageScript()', () => {
     const sampleRoot = join(__dirname, '../test-data/json/package');
@@ -24,5 +24,11 @@ describe('Test updatePackageScript()', () => {
               },
             }
         `);
+    });
+
+    test('hasUI5CliV3', async () => {
+        expect(hasUI5CliV3({ '@ui5/cli': '3.0.0' })).toBe(true);
+        expect(hasUI5CliV3({})).toBe(false);
+        expect(hasUI5CliV3({ '@ui5/cli': '2.0.0' })).toBe(false);
     });
 });
